login and registration now have their own routes.
