+++ src/components/Profile.js
@@ -5,14 +5,13 @@ const Profile = ({ userImg, profileText, thisUser, token }) => {
   
   const [expAvatar, setExpAvatar] = useState(false)
   const [expBio, setExpBio] = useState(false)
-  const [errors, setErrors] = useState(null)
   const [avatar, setAvatar] = useState('')
-  const [bio, setBio] = useState('')
+  const [bio, setBio] = useState(profileText)
 
   const handleAvatarSubmit = (event) => {
     event.preventDefault()
     console.log("click happens")
-    console.log(token)
+    console.log(avatar)
     axios.patch("https://questions-t10.herokuapp.com/auth/users/me/", {
       "image_url": avatar,
     }, {
@@ -23,8 +22,9 @@ const Profile = ({ userImg, profileText, thisUser, token }) => {
     })
       .then((data) => {
         console.log(data)
+        setAvatar(avatar)
       })
-      .catch((error) => setErrors(error.message))
+      .catch((error) => alert(error.message))
   }
 
   const handleBioSubmit = (event) => {
@@ -39,9 +39,9 @@ const Profile = ({ userImg, profileText, thisUser, token }) => {
       }
     })
       .then((data) => {
-        console.log(data)
+        setBio(bio)
       })
-      .catch((error) => setErrors(error.message))
+      .catch((error) => alert(error.message))
   }
 
   return (
@@ -51,8 +51,6 @@ const Profile = ({ userImg, profileText, thisUser, token }) => {
         <button onClick={() => (expAvatar ? setExpAvatar(false) : setExpAvatar(true))}>Edit Avatar</button>
         {expAvatar && (
           <form onSubmit={handleAvatarSubmit}>
-          {/* conditionally show error message */}
-          {errors && <div className="bg-red white pa3">{errors}</div>}
           <div className="mv2">
               <label className="db mb2" htmlFor="avatar">
                 Link to New Avatar Image
@@ -69,12 +67,10 @@ const Profile = ({ userImg, profileText, thisUser, token }) => {
         )}
       </div>
       <div>
-        <p>{profileText}</p>
+        <p>{bio}</p>
         <button onClick={() => (expBio ? setExpBio(false) : setExpBio(true))}>Edit Bio</button>
         {expBio && (
           <form onSubmit={handleBioSubmit}>
-          {/* conditionally show error message */}
-          {errors && <div className="bg-red white pa3">{errors}</div>}
           <div className="mv2">
               <label className="db mb2" htmlFor="username">
                 Edit Bio
